+++ CHANGELOG.md
@@ -6,6 +6,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Neotest: Basic support for `require('neotest').run.run { suite = true }`.
+  This will run the current crate's test suite, if detected.
+  Note that positions are still only discovered for buffers with an attached
+  LSP client.
+
 ## [4.4.0] - 2024-02-01
 
 ### Added